chore(bonde-styleguide): change Navbar to Header
