+++ packages/bonde-styleguide/src/index.js
@@ -8,4 +8,4 @@ WebFont.load({
 
 export { default as Button } from './Button'
 export { default as Title } from './Title'
-export { default as Navbar } from './Navbar'
+export { default as Header } from './Header'
